\name{noeffect}
\alias{noeffect}
\docType{data}
\title{Noeffect Data}
\description{
An example dataset to illustrate how a non functional consequence may appear in a one person observation stream.
}
\usage{data("noeffect")}

\format{
   A data frame with 11 observations (10 behaviors) on the following 3 variables.

  \describe{
    \item{\code{VIDELT}}{a numeric vector indicating the observational episode}
    \item{\code{TAR}}{a character vector indicating the actor being observed}
    \item{\code{BEH}}{a character vector indicating the behavior is target ('o'), reinforcer ('A') or behavioral non-target ('x')}
}
}

